+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.3]
+### Added
+- Client can read input from stdin.
+### Fixed
+- Work around a non-understood error.
+
 ## [v0.1.2]
 ### Fixed
 - Batching of reads for gRPC stream was incorrect, leading to client errors.